backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-generativeai==0.8.3
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.18
cryptography==44.0.0
